UI compatibility and new features
1. added compatibility with different iSO devices, iPhone, iPad, both
portrait and landscape

2. added push notification when time is about to expire to remind user
to add time or check out.

3. added push notification when time expired to remind the user that
he/she has been checked out automatically.

4. realized the "adding time" feature when time is about to expire.
Specifically, when there are two minutes left, the app brings up an
alert sheet if the app is running in the foreground, or pops up push
notification if the app is running in the background.

5. refined UI, including icons and font sizes